BMW Renaming + CAN wake and Sleep
E31 - Wake and sleep CAN on T15
E39 - Wake and Sleep CAN on ABS 0x1F3
E65+ E9x - Wake and Sleep CAN on Network mng 0x480
