+++ include/BMW_E39.h
@@ -16,8 +16,8 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-#ifndef Can_E39_h
-#define Can_E39_h
+#ifndef BMW_E39_h
+#define BMW_E39_h
 
 /*  This library supports the Powertrain CAN messages for the BMW E39 for driving dash gauges, putting out malf lights etc
 
@@ -27,12 +27,13 @@
 #include "vehicle.h"
 
 
-class Can_E39: public Vehicle
+class BMW_E39: public Vehicle
 {
 
 public:
    void SetCanInterface(CanHardware* c);
    void Task10Ms();
+   void Task100Ms();
    void SetRevCounter(int s) { speed = s; }
    void SetTemperatureGauge(float temp);
    void DecodeCAN(int id, uint32_t* data);
@@ -49,6 +50,8 @@ private:
 
    uint16_t speed;
    bool isE46;
+   bool AbsCANalive;
+   bool SendCAN;
 };
 
 #endif /* Can_E39_h */
