+++ src/BMW_E31.cpp
@@ -16,10 +16,10 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
- #include "BMW_E31.h"
- #include "hwinit.h"
- #include <libopencm3/stm32/timer.h>
- #include <libopencm3/stm32/gpio.h>
+#include "BMW_E31.h"
+#include "hwinit.h"
+#include <libopencm3/stm32/timer.h>
+#include <libopencm3/stm32/gpio.h>
 /*
 *E31 840CI Tacho:
 *1000RPM = 70Hz
@@ -30,166 +30,169 @@
 
 
 //We use this as an init function
-void Bmw_E31::SetCanInterface(CanHardware* c)
+void BMW_E31::SetCanInterface(CanHardware* c)
 {
-   can = c;
-   tim_setup();//Fire up timer one...
-   timer_disable_counter(TIM1);//...but disable until needed
-   //note we are trying to reuse the lexus gs450h oil pump pwm output here to drive the tach
-   //Be aware this will prevent combo of E31 and GS450H for now ...
-   timerIsRunning=false;
+    can = c;
+    tim_setup();//Fire up timer one...
+    timer_disable_counter(TIM1);//...but disable until needed
+    //note we are trying to reuse the lexus gs450h oil pump pwm output here to drive the tach
+    //Be aware this will prevent combo of E31 and GS450H for now ...
+    timerIsRunning=false;
 
-   can->RegisterUserMessage(0x153);//ASC message. Will confirm.
+    can->RegisterUserMessage(0x153);//ASC message. Will confirm.
 }
 
 
-void Bmw_E31::SetRevCounter(int speed)
+void BMW_E31::SetRevCounter(int speed)
 {
 
-   uint16_t speed_input = speed;
-   speed_input = MAX(750, speed_input);//
-   speed_input = MIN(7500, speed_input);
-   timerPeriod = 30000000 / speed_input; //TODO: find correct factor or make parameter. current gives 52Hz at 750rpm.
-   timer_set_period(TIM1, timerPeriod);
-   timer_set_oc_value(TIM1, TIM_OC1, timerPeriod / 2); //always stay at 50% duty cycle
+    uint16_t speed_input = speed;
+    speed_input = MAX(750, speed_input);//
+    speed_input = MIN(7500, speed_input);
+    timerPeriod = 30000000 / speed_input; //TODO: find correct factor or make parameter. current gives 52Hz at 750rpm.
+    timer_set_period(TIM1, timerPeriod);
+    timer_set_oc_value(TIM1, TIM_OC1, timerPeriod / 2); //always stay at 50% duty cycle
 
 }
 
 
-void Bmw_E31::SetTemperatureGauge(float temp)
+void BMW_E31::SetTemperatureGauge(float temp)
 {
-   float dc = temp * 10; //TODO find right factor for value like 0..0.5 or so
-   //Would like to use digi pots here
-   dc = dc;
+    float dc = temp * 10; //TODO find right factor for value like 0..0.5 or so
+    //Would like to use digi pots here
+    dc = dc;
 }
 
-void Bmw_E31::DecodeCAN(int id, uint32_t* data)
+void BMW_E31::DecodeCAN(int id, uint32_t* data)
 {
-   uint8_t* bytes = (uint8_t*)data;//E31 CAN to be added here
+    uint8_t* bytes = (uint8_t*)data;//E31 CAN to be added here
 
-     if (id == 0x153)// ASC1 contains road speed signal. Unsure if applies to E31 as yet ....
+    if (id == 0x153)// ASC1 contains road speed signal. Unsure if applies to E31 as yet ....
     {
-      //Vehicle speed signal in Km/h
-      //Calculation = ( (HEX[MSB] * 256) + HEX[LSB]) * 0.0625
-      //Min: 0x160 (0 Km/h)
+        //Vehicle speed signal in Km/h
+        //Calculation = ( (HEX[MSB] * 256) + HEX[LSB]) * 0.0625
+        //Min: 0x160 (0 Km/h)
 
-      float road_speed = 0.0625f * (((bytes[2] << 8) | (bytes[1])) - 0x160);
+        float road_speed = 0.0625f * (((bytes[2] << 8) | (bytes[1])) - 0x160);
 
-      Param::SetFloat(Param::Veh_Speed, road_speed);
+        Param::SetFloat(Param::Veh_Speed, road_speed);
     }
 }
 
-void Bmw_E31::EGSMsg43B()  //EGS1
+void BMW_E31::EGSMsg43B()  //EGS1
 {
 
-   uint8_t bytes[3];
+    uint8_t bytes[3];
 
-   bytes[0]=0x46;
-   bytes[1]=0x00;
-   bytes[2]=0x00;
+    bytes[0]=0x46;
+    bytes[1]=0x00;
+    bytes[2]=0x00;
 
-   can->Send(0x43B, (uint32_t*)bytes,3);
+    can->Send(0x43B, (uint32_t*)bytes,3);
 }
 
-void Bmw_E31::EGSMsg43F(int8_t gear)
+void BMW_E31::EGSMsg43F(int8_t gear)
 {
-   //Can bus data packet values to be sent
-   uint8_t bytes[8];
-   // Source: https://www.bimmerforums.com/forum/showthread.php?1887229-E46-Can-bus-project&p=30055342#post30055342
-   // byte 0 = 0x81 //doesn't do anything to the ike
-   bytes[0] = 0x81;
-   // byte 1 = 0x01 where;
-   // 01 = first gear
-   // 02= second gear
-   // 03 = third gear
-   // 04 = fourth gear
-   // 05 = D
-   // 06 = N
-   // 07 = R
-   // 08 = P
-   // 09 = 5
-   // 0A = 6
-   switch (gear)
-   {
-   case -1 /* Reverse */:
-      bytes[1] = 0x07;
-      break;
-   case 0 /* Neutral */:
-      bytes[1] = 0x06;
-      break;
-   case 1 /* Drive */:
-      bytes[1] = 0x05;
-      break;
-   default:
-      bytes[1] = 0x08;
-      break;
-   }
+    //Can bus data packet values to be sent
+    uint8_t bytes[8];
+    // Source: https://www.bimmerforums.com/forum/showthread.php?1887229-E46-Can-bus-project&p=30055342#post30055342
+    // byte 0 = 0x81 //doesn't do anything to the ike
+    bytes[0] = 0x81;
+    // byte 1 = 0x01 where;
+    // 01 = first gear
+    // 02= second gear
+    // 03 = third gear
+    // 04 = fourth gear
+    // 05 = D
+    // 06 = N
+    // 07 = R
+    // 08 = P
+    // 09 = 5
+    // 0A = 6
+    switch (gear)
+    {
+    case -1 /* Reverse */:
+        bytes[1] = 0x07;
+        break;
+    case 0 /* Neutral */:
+        bytes[1] = 0x06;
+        break;
+    case 1 /* Drive */:
+        bytes[1] = 0x05;
+        break;
+    default:
+        bytes[1] = 0x08;
+        break;
+    }
 
-   // byte 2 = 0xFF where;
-   // FF = no display
-   // 00 = E
-   // 39 = M
-   // 40 = S
-   bytes[2] = 0xFF;
+    // byte 2 = 0xFF where;
+    // FF = no display
+    // 00 = E
+    // 39 = M
+    // 40 = S
+    bytes[2] = 0xFF;
 
-   // byte 3 = 0xFF //doesn't do anything to the ike
-   bytes[3] = 0xFF;
+    // byte 3 = 0xFF //doesn't do anything to the ike
+    bytes[3] = 0xFF;
 
-   // byte 4 = 0x00 //doesn't do anything to the ike
-   bytes[4] = 0x00;
+    // byte 4 = 0x00 //doesn't do anything to the ike
+    bytes[4] = 0x00;
 
-   // byte 5 = 0x80 where;
-   // 80 = clears the gear warning picture - all other values bring it on
-   bytes[5] = 0x80;
+    // byte 5 = 0x80 where;
+    // 80 = clears the gear warning picture - all other values bring it on
+    bytes[5] = 0x80;
 
-   // byte 6 = 0xFF //doesn't do anything to the ike
-   bytes[6] = 0xFF;
+    // byte 6 = 0xFF //doesn't do anything to the ike
+    bytes[6] = 0xFF;
 
-   // byte 7 = 0x00 //doesn't do anything to the ike
-   bytes[7] = 0xFF;
+    // byte 7 = 0x00 //doesn't do anything to the ike
+    bytes[7] = 0xFF;
 
-   can->Send(0x43F, bytes, 8);
+    can->Send(0x43F, bytes, 8);
 }
 
- void Bmw_E31::Task1Ms()
+void BMW_E31::Task1Ms()
 {
 
-
-}
-
-
- void Bmw_E31::Task10Ms()
-{
-   EGSMsg43B();//EGS1
-   if (Param::GetBool(Param::Transmission))
-      EGSMsg43F(Param::GetInt(Param::dir));
 
 }
 
 
-void Bmw_E31::Task100Ms()
+void BMW_E31::Task10Ms()
 {
-
-   if(Param::GetInt(Param::T15Stat) && !timerIsRunning)
-   {
-   timer_enable_counter(TIM1);
-   timerIsRunning=true;
-   }
-   else if (!Param::GetInt(Param::T15Stat) && timerIsRunning)
-   {
-   timer_disable_counter(TIM1);
-   timerIsRunning=false;
-   }
+    if(Param::GetInt(Param::T15Stat))
+    {
+        EGSMsg43B();//EGS1
+        if (Param::GetBool(Param::Transmission))
+            EGSMsg43F(Param::GetInt(Param::dir));
+    }
 
 }
 
 
- bool Bmw_E31::Ready()
+void BMW_E31::Task100Ms()
 {
-   return DigIo::t15_digi.Get();
+
+    if(Param::GetInt(Param::T15Stat) && !timerIsRunning)
+    {
+        timer_enable_counter(TIM1);
+        timerIsRunning=true;
+    }
+    else if (!Param::GetInt(Param::T15Stat) && timerIsRunning)
+    {
+        timer_disable_counter(TIM1);
+        timerIsRunning=false;
+    }
+
 }
 
-bool Bmw_E31::Start()
+
+bool BMW_E31::Ready()
 {
-   return Param::GetBool(Param::din_start);
+    return DigIo::t15_digi.Get();
+}
+
+bool BMW_E31::Start()
+{
+    return Param::GetBool(Param::din_start);
 }
